Task 관련 작성

+++ android/app/src/main/java/com/downfall/caterplanner/common/model/BriefingStatizable.java
@@ -6,8 +6,8 @@ import java.util.Date;
 
 public interface BriefingStatizable {
 
-    float progress();
-    float achieve();
+    int progress();
+    int achieve();
 
     void statistion();
